requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.16
langchain-text-splitters==0.2.16
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
arabert==1.0.0
ar-corrector==1.1.6
pyarabic==0.6.15
pypdf==3.17.1
python-docx==1.0.1
fastapi[all]==0.110.0
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.7
streamlit==1.33.0
nltk==3.8.1
mistralai==0.1.5
python-dotenv==1.0.0
chromadb==0.4.22